+++ src/pages/FriendPage/FriendChatRoomPage/components/chat-friend-list/styles.css.ts
@@ -16,6 +16,7 @@ export const friendList = style({
   overflowY: 'auto',
   overflowX: 'hidden',
 });
+
 export const list = style({
   listStyle: 'none',
   padding: 0,
@@ -23,6 +24,7 @@ export const list = style({
   display: 'flex',
   flexDirection: 'column',
 });
+
 export const item = style({
   display: 'flex',
   alignItems: 'center',
@@ -38,12 +40,14 @@ export const item = style({
     backgroundColor: theme.color.grayHover,
   },
 });
+
 export const avatar = style({
   width: rem(42),
   height: rem(42),
   borderRadius: '50%',
   objectFit: 'cover',
 });
+
 export const divider = style({
   width: rem(200),
   height: rem(1),
