+++ src/pages/FriendPage/FriendChatRoomPage/index.tsx
@@ -2,8 +2,8 @@ import { useNavigate } from 'react-router-dom';
 
 import { ChatBox } from './components/chat-box';
 import { FriendHeader } from './components/chat-friend-header';
-// import { FriendList } from './components/chat-friend-list';
-import * as styles from './FriendChatRoomPage.styles.css';
+import { FriendList } from './components/chat-friend-list';
+import * as styles from './styles.css.ts';
 
 export const FriendChatRoomPage = () => {
   const navigate = useNavigate();
@@ -22,7 +22,9 @@ export const FriendChatRoomPage = () => {
             onClick={() => navigate('/friend')}
           />
         </div>
-        <div className={styles.friendListWrapper}>{/* <FriendList /> */}</div>
+        <div className={styles.friendListWrapper}>
+          <FriendList />
+        </div>
       </div>
     </div>
   );
